pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "nltk>=3.9.1",
    "openai>=1.65.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "weaviate>=0.1.2",
]
